feat: Improve TUI mouse support
The TUI currently has opt-in mouse capability (via `-mouse` flag) due to limited support and some bugs.

This commit resolves the issues with mouse support by forcing the TUI to render in an alternative screen buffer. As such, the `-mouse` flag has been replaced with a `-no-mouse` flag in a change from opt-in to opt-out.